Teach the object oriented matplotlib API instead.
So far in course "202 - Plots mit Matplotlib" the procedural API of
matplotlib was taught. This commit changes the course, to teach the
object oriented API instead, which only adds additional complexity in
the form of 2 lines of boilerplate code per figure, but adds a lot of
flexibility. After all, the ooo API is unavoidable, even this course
already used it for the multiple plots section, so it seems logical to
teach it from the beginning.